[REF] Factor out data dictionary code
Create the dictionary separately in a utility
function.
