+++ src/components/Download.tsx
@@ -11,6 +11,7 @@ import {
 } from '@mui/material';
 import Ajv from 'ajv';
 import { useState, useMemo } from 'react';
+import { getDataDictionary } from '~/utils/util';
 import emoji from '../assets/download-emoji.png';
 import schema from '../assets/neurobagel_data_dictionary.schema.json';
 import useDataStore from '../stores/data';
@@ -18,102 +19,20 @@ import useViewStore from '../stores/view';
 import { DataDictionary, View } from '../utils/internal_types';
 import DataDictionaryPreview from './DataDictionaryPreview';
 
+function useDataDictionary(): DataDictionary {
+  const columns = useDataStore((state) => state.columns);
+  return useMemo(() => getDataDictionary(columns), [columns]);
+}
+
 function Download() {
   const [dictionaryCollapsed, setDictionaryCollapsed] = useState(false);
   const [forceAllowDownload, setForceAllowDownload] = useState(false);
 
   const uploadedDataTableFileName = useDataStore((state) => state.uploadedDataTableFileName);
-  const columns = useDataStore((state) => state.columns);
+  // const columns = useDataStore((state) => state.columns);
   const reset = useDataStore((state) => state.reset);
-
+  const dataDictionary: DataDictionary = useDataDictionary();
   const setCurrentView = useViewStore((state) => state.setCurrentView);
-
-  const dataDictionary = useMemo(
-    () =>
-      Object.entries(columns).reduce((dictAcc, [_columnKey, column]) => {
-        if (column.header) {
-          const dictionaryEntry: DataDictionary[string] = {
-            Description: column.description || '',
-          };
-
-          // Description of levels always included for the BIDS section
-          if (column.dataType === 'Categorical' && column.levels) {
-            dictionaryEntry.Levels = Object.entries(column.levels).reduce(
-              (levelsObj, [levelKey, levelValue]) => ({
-                ...levelsObj,
-                [levelKey]: {
-                  Description: levelValue.description || '',
-                },
-              }),
-              {} as { [key: string]: { Description: string } }
-            );
-          }
-
-          if (column.dataType === 'Continuous' && column.units !== undefined) {
-            dictionaryEntry.Units = column.units;
-          }
-
-          if (column.standardizedVariable) {
-            dictionaryEntry.Annotations = {
-              IsAbout: {
-                TermURL: column.standardizedVariable.identifier,
-                Label: column.standardizedVariable.label,
-              },
-            };
-
-            // Add term url to Levels under BIDS section only for a categorical column with a standardized variable
-            if (column.dataType === 'Categorical' && column.levels) {
-              dictionaryEntry.Levels = Object.entries(column.levels).reduce(
-                (updatedLevels, [levelKey, levelValue]) => ({
-                  ...updatedLevels,
-                  [levelKey]: {
-                    ...updatedLevels[levelKey],
-                    ...(levelValue.termURL ? { TermURL: levelValue.termURL } : {}),
-                  },
-                }),
-                dictionaryEntry.Levels || {}
-              );
-
-              dictionaryEntry.Annotations.Levels = Object.entries(column.levels).reduce(
-                (termsObj, [levelKey, levelValue]) => ({
-                  ...termsObj,
-                  [levelKey]: {
-                    TermURL: levelValue.termURL || '',
-                    Label: levelValue.label || '',
-                  },
-                }),
-                {} as { [key: string]: { TermURL: string; Label: string } }
-              );
-            }
-
-            if (column.isPartOf?.termURL && column.isPartOf?.label) {
-              dictionaryEntry.Annotations.IsPartOf = {
-                TermURL: column.isPartOf.termURL,
-                Label: column.isPartOf.label,
-              };
-            }
-
-            if (column.missingValues && column.dataType !== null) {
-              dictionaryEntry.Annotations.MissingValues = column.missingValues;
-            }
-
-            if (column.dataType === 'Continuous' && column.format) {
-              dictionaryEntry.Annotations.Format = {
-                TermURL: column.format?.termURL || '',
-                Label: column.format?.label || '',
-              };
-            }
-          }
-
-          return {
-            ...dictAcc,
-            [column.header]: dictionaryEntry,
-          };
-        }
-        return dictAcc;
-      }, {} as DataDictionary),
-    [columns]
-  );
 
   const { isValid: schemaValid, errors: schemaErrors } = useMemo(() => {
     const ajv = new Ajv({ allErrors: true });
